+++ android/app/src/main/AndroidManifest.xml
@@ -15,17 +15,16 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:label="@string/title_activity_main" >
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
-            android:name=".Result"
+            android:name=".activities.Result"
             android:label="@string/title_activity_result" >
         </activity>
 
@@ -37,14 +36,12 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".MapsActivity"
+            android:name=".activities.MapsActivity"
             android:label="@string/title_activity_maps" >
         </activity>
 
-=======
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="" />
->>>>>>> master
     </application>
 </manifest>
