+++ android/app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="com.osu.cse.projectblocks.MainActivity">
+      tools:context="com.osu.cse.projectblocks.activities.MainActivity">
       <item
           android:id="@+id/find_food"
           android:orderInCategory="100"
